requirements.txt
httpx>=0.25
python-dotenv>=1.0
uagents>=0.11
uagents-adapter>=0.1.6
fastmcp>=2.13
